fix: consolidate functionality used for adding third party scripts
