+++ components/shared/publizonPlayer/helper.ts
@@ -1,7 +1,4 @@
-type TAssetType = {
-  src: string
-  type: "script" | "link"
-}
+import { TAssetType } from "@/lib/helpers/helper.scripts"
 
 export const assets: TAssetType[] = [
   {
@@ -9,21 +6,3 @@ export const assets: TAssetType[] = [
     type: "script",
   },
 ]
-
-export const appendAsset = ({ src, type }: TAssetType) => {
-  if (type === "script") {
-    const scriptElement = document.createElement("script")
-    scriptElement.src = src
-    scriptElement.defer = true
-    scriptElement.async = false
-    scriptElement.type = "module"
-    document.head.appendChild(scriptElement)
-  }
-}
-
-export const removeAsset = ({ src, type }: TAssetType) => {
-  if (type === "script") {
-    const scriptElement = document.querySelector(`script[src="${src}"]`)
-    scriptElement?.remove()
-  }
-}
