+++ components/shared/publizonReader/PublizonReader.tsx
@@ -2,7 +2,9 @@
 
 import React, { useEffect } from "react"
 
-import { appendAsset, readerAssets, removeAsset } from "./helper"
+import { appendAsset, removeAsset } from "@/lib/helpers/helper.scripts"
+
+import { readerAssets } from "./helper"
 
 // Define mutually exclusive types for identifier and orderId
 type ReaderType =
